+++ src/Plugins.Win32.TASRSP/Main.cpp
@@ -79,7 +79,7 @@ void handle_unknown_task(const OSTask_t *task, const uint32_t sum)
         f = fopen("disasm.txt", "wb");
         memcpy(rsp.dmem, rsp.rdram + task->ucode_data, task->ucode_data_size);
         memcpy(rsp.imem + 0x80, rsp.rdram + task->ucode, 0xF7F);
-        disasm(f, (unsigned long *)(rsp.imem));
+        disasm(f, (uint32_t *)(rsp.imem));
         fclose(f);
     }
     else
@@ -93,7 +93,7 @@ void handle_unknown_task(const OSTask_t *task, const uint32_t sum)
         fclose(f);
 
         f = fopen("disasm.txt", "wb");
-        disasm(f, (unsigned long *)(rsp.imem));
+        disasm(f, (uint32_t *)(rsp.imem));
         fclose(f);
     }
 }
@@ -115,13 +115,13 @@ void audio_ucode_zelda()
 
 int audio_ucode_detect_type(const OSTask_t *task)
 {
-    if (*(unsigned long *)(rsp.rdram + task->ucode_data + 0) != 0x1)
+    if (*(uint32_t *)(rsp.rdram + task->ucode_data + 0) != 0x1)
     {
         if (*(rsp.rdram + task->ucode_data + (0 ^ 3 - S8)) == 0xF) return 4;
         return 3;
     }
 
-    if (*(unsigned long *)(rsp.rdram + task->ucode_data + 0x30) == 0xF0000F00) return 1;
+    if (*(uint32_t *)(rsp.rdram + task->ucode_data + 0x30) == 0xF0000F00) return 1;
     return 2;
 }
 
@@ -178,9 +178,9 @@ int audio_ucode(OSTask_t *task)
 
     g_audio_ucode_func();
 
-    const auto p_alist = (unsigned long *)(rsp.rdram + task->data_ptr);
+    const auto p_alist = (uint32_t *)(rsp.rdram + task->data_ptr);
 
-    for (unsigned int i = 0; i < task->data_size / 4; i += 2)
+    for (uint32_t i = 0; i < task->data_size / 4; i += 2)
     {
         inst1 = p_alist[i];
         inst2 = p_alist[i + 1];
@@ -207,7 +207,7 @@ void on_rom_closed()
 uint32_t do_rsp_cycles(uint32_t Cycles)
 {
     OSTask_t *task = (OSTask_t *)(rsp.dmem + 0xFC0);
-    unsigned int i, sum = 0;
+    uint32_t i, sum = 0;
 
     g_rsp_alive = true;
